refactor: LinkedList 파일 분리 #2
